+++ ts/test/brrr.ts
@@ -1,10 +1,7 @@
 import { deepStrictEqual, rejects } from 'node:assert'
 import { suite, test } from 'node:test'
-import { Brrr } from '../src'
-import { InMemoryStore } from '../src/adapters/in-memory-store'
-import { Memory } from '../src/models/memory'
+import { Brrr, InMemoryStore, Memory, NaiveCodec } from '../src'
 import { ClosableInMemQueue } from './fixtures/closable-mem-queue'
-import { NaiveCodec } from './fixtures/naive-codec'
 
 await suite(import.meta.filename, async () => {
   await suite('No Brrr', async () => {
